:lipstick: Style: 회원가입 스텝 보여주는 아이콘 추가 및 디자인
